requirements.txt
mcap==1.1.1
mcap-ros2-support==0.5.3